COG-1197: AuditService is extracted
